PLATFORM/backend/api/pyproject.toml
[tool.poetry]
name = "bracelet-api"
version = "0.1.0"
description = "Virtual Experience HTTP REST API"
license = "Proprietary"
authors = ["Aprendiendo <depablodiegouma@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.5.8"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"
antlr4-python3-runtime = "^4.9.3"
python-dotenv = "^0.20.0"
PyJWT = "^2.1.0"
python-multipart = "^0.0.5"
Mako = "^1.1.6"
Faker = "^13.0.0"
async-cache = "^1.1.1"
pytz = "^2022.6"
boto3 = "^1.26.58"
pyedflib = "^0.1.37"
six = "^1.17.0"

[tool.poetry.dev-dependencies]
bracelet-lib = {path = "../bracelet-lib", develop=true}
pytest = "^5.2"
pydevd-pycharm = "^221.4165.171"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
